Add genesis data
This includes the genesis data for the mainnet, testnet and staging
chains in the executable, allowing 'cardano-cli blockchain verify' to
work out of the box. In the future, since including this stuff adds
quite a bit of bloat to the binaries, we may want to fetch genesis
data from Hermes (though that doesn't help if we're using cardano-cli
with native peers).