update the input selection algorithm for the command `input-select`
this comes as we needed to update the cardano-deps. The new input
selection schemes is more precise and provides more algorithms.

Only 2 are really exposed to the command line users:

* largest first;
* blackjack.

It is because only these 2 makes sense at the usage as they have
meaning. It is possible the blackjack algorithm fails with not
enough inputs. The user will need to fallback on the largest first
algorithm in that case