wallet: Improve LogLock
It's probably not safe to have the client pass back the log file path to
delete, expecting it to be the same as the locked log file.
Better to remember the path inside LogLock.
Shortened the method names and named the log deletion method
delete_wallet_log because it's the log file that is being deleted.