fix dialoguer lost commit (this is due to a rebase of the PR)
I made a PR to the upstream repo and had to rebase it. The commit
we depends on has disapeared and I needed to upgrade the CLI to
compile it again.

There is a regression though in the sense I need to change the theme
setting to erase the line when prompting the mnemonics or when
displaying the mnemoncis words at creating time